get rid of some compiler warnings

+++ FuzzyKClusters.cpp
@@ -49,9 +49,9 @@ MatrixXd KMeans::Initialize(const MatrixXd &data, int k, double q)
 
     // guess the cluster centers
     mems = MatrixXd::Random(k, data.rows());
-    for (size_t j = 0; j < mems.cols(); ++j) {
+    for (int j = 0; j < mems.cols(); ++j) {
         auto csum = mems.col(j).sum();
-        for (size_t i = 0; i < mems.rows(); ++i) {
+        for (int i = 0; i < mems.rows(); ++i) {
             mems(i, j) = mems(i, j)/csum;
         }
     }
@@ -64,8 +64,8 @@ MatrixXd KMeans::Initialize(const MatrixXd &data, int k, double q)
 // distance matrix (num_clusters, num_data)
 void KMeans::Distances(const MatrixXd &centroids, const MatrixXd &data)
 {
-    for (size_t i = 0; i < centroids.rows(); ++i) {
-        for (size_t j = 0; j < data.rows(); ++j) {
+    for (int i = 0; i < centroids.rows(); ++i) {
+        for (int j = 0; j < data.rows(); ++j) {
             dists(i, j) = (centroids.row(i) - data.row(j)).cwiseAbs2().sum();
         }
     }
@@ -77,9 +77,9 @@ void KMeans::Memberships(double q)
     // coeffcient-wise operation
     auto d = dists.array().pow(-1.0/(q - 1.0)).matrix();
 
-    for (size_t j = 0; j < d.cols(); ++j) {
+    for (int j = 0; j < d.cols(); ++j) {
         auto dsum = d.col(j).sum();
-        for (size_t i = 0; i < d.rows(); ++i) {
+        for (int i = 0; i < d.rows(); ++i) {
             mems(i, j) = d(i, j)/dsum;
         }
     }
@@ -89,9 +89,9 @@ void KMeans::Memberships(double q)
 void KMeans::FormClusters(MatrixXd &clusters, const MatrixXd &data, double q)
 {
     auto weights = mems.array().pow(q).matrix();
-    for (size_t i = 0; i < clusters.rows(); ++i) {
+    for (int i = 0; i < clusters.rows(); ++i) {
         clusters.row(i) *= 0;
-        for (size_t j = 0; j < data.rows(); ++j) {
+        for (int j = 0; j < data.rows(); ++j) {
             clusters.row(i) += data.row(j)*weights(i, j);
         }
         clusters.row(i) /= weights.row(i).sum();
@@ -155,8 +155,8 @@ void KRings::Distances(const MatrixXd &centroids, const MatrixXd &data)
     auto const centers = centroids.leftCols(centroids.cols() - 1);
     auto const radii = centroids.rightCols(1);
 
-    for (size_t i = 0; i < centroids.rows(); ++i) {
-        for (size_t j = 0; j < data.rows(); ++j) {
+    for (int i = 0; i < centroids.rows(); ++i) {
+        for (int j = 0; j < data.rows(); ++j) {
             dists_euc(i, j) = std::sqrt((centers.row(i) - data.row(j)).cwiseAbs2().sum());
             dists(i, j) = std::pow(dists_euc(i, j) - radii(i, 0), 2);
         }
@@ -169,9 +169,9 @@ void KRings::FormRadii(MatrixXd &clusters, double q)
     auto radii = clusters.rightCols(1);
     auto weights = mems.array().pow(q).matrix();
 
-    for (size_t i = 0; i < weights.rows(); ++i) {
+    for (int i = 0; i < weights.rows(); ++i) {
         radii(i, 0) = 0;
-        for (size_t j = 0; j < weights.cols(); ++j) {
+        for (int j = 0; j < weights.cols(); ++j) {
             radii(i, 0) += weights(i, j)*dists_euc(i, j);
         }
         radii(i, 0) /= weights.row(i).sum();
@@ -185,10 +185,10 @@ void KRings::FormClusters(MatrixXd &clusters, const MatrixXd &data, double q)
     const auto &radii = clusters.rightCols(1);
     auto weights = mems.array().pow(q).matrix();
 
-    for (size_t i = 0; i < weights.rows(); ++i) {
+    for (int i = 0; i < weights.rows(); ++i) {
         MatrixXd icenter = centers.row(i);
         centers.row(i) *= 0;
-        for (size_t j = 0; j < weights.cols(); ++j) {
+        for (int j = 0; j < weights.cols(); ++j) {
             double scale = radii(i, 0)/dists_euc(i, j);
             centers.row(i) += weights(i, j)*(data.row(j) - (data.row(j) - icenter)*scale);
         }
